Portfolio/IBKR_Trading_Engine/FE/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

jinja2==3.1.3

python-dotenv==1.0.1

wheel

black

httpx==0.27.0

python-multipart==0.0.9

watchfiles

pre-commit

ldap3==2.9.1

itsdangerous==2.2.0

pylint==3.3.1